+++ app/test/service/email/backend_test.dart
@@ -5,6 +5,7 @@
 import 'package:clock/clock.dart';
 import 'package:pub_dev/fake/backend/fake_email_sender.dart';
 import 'package:pub_dev/service/email/backend.dart';
+import 'package:pub_dev/service/email/models.dart';
 import 'package:pub_dev/shared/datastore.dart';
 import 'package:pub_dev/shared/email.dart';
 import 'package:test/test.dart';
@@ -16,7 +17,7 @@ void main() {
     testWithProfile('one immediate success + one success on second attempt',
         fn: () async {
       fakeEmailSender.failNextMessageCount = 1;
-      final entries = emailBackend.prepareEntities(
+      final entry = emailBackend.prepareEntity(
         EmailMessage(
           EmailAddress('from@pub.dev'),
           [
@@ -27,8 +28,8 @@ void main() {
           'body',
         ),
       );
-      await dbService.commit(inserts: entries);
-      await emailBackend.trySendOutgoingEmails(entries);
+      await dbService.commit(inserts: [entry]);
+      await emailBackend.trySendOutgoingEmail(entry);
       expect(fakeEmailSender.sentMessages, hasLength(1));
       final email = fakeEmailSender.sentMessages.single;
       expect(email.from.email, 'from@pub.dev');
@@ -55,7 +56,7 @@ void main() {
 
     testWithProfile('only failed attempts', fn: () async {
       fakeEmailSender.failNextMessageCount = 2;
-      final entries = emailBackend.prepareEntities(
+      final entry = emailBackend.prepareEntity(
         EmailMessage(
           EmailAddress('from@pub.dev'),
           [EmailAddress('to@pub.dev')],
@@ -63,7 +64,7 @@ void main() {
           'body',
         ),
       );
-      await dbService.commit(inserts: entries);
+      await dbService.commit(inserts: [entry]);
 
       for (var i = 0; i < 2; i++) {
         await withClock(Clock.fixed(clock.now().add(Duration(hours: i * 8))),
@@ -78,6 +79,48 @@ void main() {
           () async {
         expect(await emailBackend.trySendAllOutgoingEmails(), 0);
         expect(fakeEmailSender.sentMessages, isEmpty);
+        expect(await emailBackend.deleteDeadOutgoingEmails(), 1);
+      });
+    });
+
+    testWithProfile('pending claim prevents sending', fn: () async {
+      final entry = emailBackend.prepareEntity(
+        EmailMessage(
+          EmailAddress('from@pub.dev'),
+          [
+            EmailAddress('to1@pub.dev'),
+            EmailAddress('to2@pub.dev'),
+          ],
+          'subj',
+          'body',
+        ),
+      );
+      expect(entry.recipientEmails?.toSet(), {'to1@pub.dev', 'to2@pub.dev'});
+      await dbService.commit(inserts: [entry]);
+
+      OutgoingEmail? pending;
+      fakeEmailSender.failNextMessageCount = 1;
+      await withClock(Clock.fixed(clock.now().add(Duration(hours: 1))),
+          () async {
+        expect(await emailBackend.trySendAllOutgoingEmails(), 1);
+        expect(await emailBackend.deleteDeadOutgoingEmails(), 0);
+        pending = await dbService.lookupValue<OutgoingEmail>(entry.key);
+        expect(pending!.recipientEmails, hasLength(1));
+      });
+
+      // emails on a claimed entry won't be sent out
+      await withClock(Clock.fixed(clock.now().add(Duration(hours: 3))),
+          () async {
+        pending!.claimId = 'claim-uuid';
+        await dbService.commit(inserts: [pending!]);
+        expect(await emailBackend.trySendAllOutgoingEmails(), 0);
+        expect(await emailBackend.deleteDeadOutgoingEmails(), 0);
+      });
+
+      // eventually it gets deleted
+      await withClock(Clock.fixed(clock.now().add(Duration(hours: 12))),
+          () async {
+        expect(await emailBackend.trySendAllOutgoingEmails(), 0);
         expect(await emailBackend.deleteDeadOutgoingEmails(), 1);
       });
     });